+++ src/utils/api/auth/index.js
@@ -1,4 +1,4 @@
+import { loginSchema } from "./schema";
 import { login, refreshJwt } from "./api";
-import { loginSchema } from "./type";
 
 export { login, refreshJwt, loginSchema };
